feature. RDB, HIVE Service 분기 로직 1차 병합
